+++ src/kubernetes_test.go
@@ -2,14 +2,12 @@ package main
 
 import (
 	"fmt"
-	"io/ioutil"
 	"testing"
 	"time"
 
 	"github.com/newrelic/nri-kubernetes/src/apiserver"
 	"github.com/newrelic/nri-kubernetes/src/controlplane"
 	"github.com/newrelic/nri-kubernetes/src/definition"
-	"github.com/newrelic/nri-kubernetes/src/storage"
 	"github.com/sirupsen/logrus"
 	"github.com/stretchr/testify/assert"
 )
@@ -19,8 +17,6 @@ var logger = logrus.StandardLogger()
 func TestControlPlaneJobs(t *testing.T) {
 	nodeName := "ip-10.0.2.15"
 	nodeIP := "10.0.2.15"
-	tmpDir, err := ioutil.TempDir("", "test_discover")
-	assert.NoError(t, err)
 	// Setup the data returned when querying the kubelet for the pods
 	// running on the node.
 	rawGroups := definition.RawGroups{
@@ -53,9 +49,6 @@ func TestControlPlaneJobs(t *testing.T) {
 	podsFetcher := func() (definition.RawGroups, error) {
 		return rawGroups, nil
 	}
-	// Setup storage
-	store := storage.NewJSONDiskStorage(tmpDir)
-
 	// Setup the fake api server with the labels belonging to a master node.
 	apiServerClient := apiserver.TestAPIServer{
 		Mem: map[string]*apiserver.NodeInfo{
@@ -69,17 +62,16 @@ func TestControlPlaneJobs(t *testing.T) {
 	}
 
 	// When getting the control plane jobs for this node
-	cpJobs, err := controlPlaneJobs(
+	cpJobs, _ := controlPlaneJobs(
 		logger,
 		apiServerClient,
 		nodeName,
 		time.Duration(0),
-		time.Duration(0),
-		store,
 		nodeIP,
 		podsFetcher,
 		nil,
 		"test",
+		"",
 		"",
 	)
 	assert.Equal(t, len(components), len(cpJobs))